+++ NEWS.md
@@ -1,5 +1,10 @@
 # roclang 0.1.4
 
+## Bug fixes:
+
+* Change the test of a function from base package from `base::paste` function to `base::library`, since the former will result in `<NA>` when extracting the documentation for `sep` on MKL platform, as reported by CRAN.
+
+
 # roclang 0.1.3
 
 ## Bug fixes: